requirements.txt
boto3
drf-yasg
djangorestframework
